refactor(all): reorganize the files and functions into reusable packages
