+++ exec/exec_context.go
@@ -1,0 +1,15 @@
+package exec
+
+import (
+	"context"
+)
+
+type CommandExecutionContext struct {
+	Context context.Context
+}
+
+func NewCommandExecutionContext(ctx context.Context) *CommandExecutionContext {
+	return &CommandExecutionContext{
+		Context: ctx,
+	}
+}
